algo: implement the algo to minimize transactions
This is Apo's solution.
As soon as we have a  group of people whose credit sum is null,
it's sorted, the smaller and the higher are matched to
remove one at a time. That's the only difference.
Beware that the `get_tuples_with_sum` is recursive.

Improvements:
- it needs to be commented more a
- variables could be renamed
- it needs to be tested more

It's just a first draft anyway

Signed-off-by: Michael Lohier <lohiermichael@gmail.com>