api: make CRUD for events, users and expenses
For now we have the simple CRUD (Create, Read, Update and Delete)
operations for the three models: Event, User, Expense.

Note 1:
UserEventDebt model is not yet treated, as we don't necessarily want to
expose it. We just want to have it stored internally.

Note 2:
The User model needs to be refactored with securirity and other
parameters. This is just a first draft.

Signed-off-by: Michael Lohier <lohiermichael@gmail.com>